nemo_test/requirements.txt
# PhysicsNeMo PINN Requirements
# Core PhysicsNeMo packages
nvidia-physicsnemo>=1.1.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Scientific computing
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Data handling
h5py>=3.7.0

# Visualization (optional)
matplotlib>=3.5.0
plotly>=5.0.0

# Development tools (optional)
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0